// src/public-api/model/brandDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BrandDto { 
    id?: number;
    name: string;
    description?: string | null;
    logoUrl?: string | null;
    websiteUrl?: string | null;
    isActive?: boolean;
    createdAt?: string;
    productCount?: number;
}